Add files via upload
This is my very first coding project.  It is a simple board game The top left corner of the board is coordinate (0,0) and the bottom right corner is coordinate (7,7). The X piece's objective is to reach coordinate (7,7) and player o is trying to make it so player x doesn't have any legal moves. Both X piece and O pieces can only move one space at a time each term any direction. Except that O cannot move to a space that increases in both the column and row. Have fun!